\name{writeHMM}
\alias{writeHMM}
\title{
writeHMM
}
\description{
Writes an HMM to disk. hmm should be a list on the form list("pi" = ..., "A" = ..., "B" = ...).
}
\usage{
writeHMM(hmm, filename)
}
\details{
For details on the file format see http://birc.au.dk/software/zipHMM.
}
\value{
void
}
\references{
http://birc.au.dk/software/zipHMM
}
\author{
Andreas Sand <asand@birc.au.dk>
}
\seealso{}
\keyword{ hidden Markov model }
